# spec_swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: CultureMesh
host: www.culturemesh.com
basePath: /api/v1
...
securityDefinitions:
  basicAuth:
    type: basic
schemes:
  - https
paths:
  '/user/{userId}':
    get:
      tags:
        - users
      summary: Get a particular user.
      description: Gets a user.
      operationId: getUser
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/User'
        '405':
          description: Invalid input
  '/user/{userId}/networks':
    get:
      tags:
        - users
      summary: Get networks a user belongs to.
      description: ''
      operationId: getUserNetworks
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user to return.
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          type: integer
          description: |
            The number of results to return.  Between 1 and 100.
        - in: query
          name: max_registration_date
          type: string
          format: timestamp
          description: |
            The latest registration date, inclusive, in the networks returned.
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/NetworkWithJoin'
        '405':
          description: Invalid input
  '/user/{userId}/posts':
    get:
      tags:
        - users
      summary: Get list of posts made by a user.
      description: ''
      operationId: getUserPosts
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user to return.
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          type: integer
          description: |
            The number of results to return.  Between 1 and 100.
        - in: query
          name: max_id
          type: integer
          description: |
            The maximum ID, inclusive, to return data for.
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Post'
        '405':
          description: Invalid input
  '/user/{userId}/events':
    get:
      tags:
        - users
      summary: Get list of events associated with a user.
      description: ''
      operationId: getUserEvents
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of user to return.
          required: true
          type: integer
          format: int64
        - in: query
          name: role
          description: 'Can be "hosting" or " attending"  '
          required: true
          type: string
          enum:
            - hosting
            - attending
        - in: query
          name: count
          type: integer
          description: |
            The number of results to return.  Between 1 and 100.
        - in: query
          name: max_id
          type: integer
          description: |
            The maximum ID, inclusive, to return data for.
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Event'
        '405':
          description: Invalid input
...
  '/user/joinEvent/{eventId}':
    post:
      security:
            - basicAuth: []
      tags:
        - users
      summary: Add user to an event.
      description: ''
      operationId: addUserToEvent
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path
          description: ID of event to add user to.
          required: true
          type: integer
        - name: role
          in: query
          description: role of user. Either 'host' or 'guest'
          required: true
          type: string
      responses:
        '200':
          description: OK
        '405':
          description: Invalid input
  '/user/leaveEvent/{eventId}':
    delete:
      security:
            - basicAuth: []
      tags:
        - users
      summary: Removes a user from an event.
      description: ''
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path
          description: ID of event to add user to.
          required: true
          type: integer
      responses:
        '200':
          description: OK
        '405':
          description: Invalid input
  '/user/joinNetwork/{networkId}':
    post:
      security:
        - basicAuth: []
      tags:
        - users
      summary: Adds user to a network.
      description: ''
      operationId: addUserToNetwork
      produces:
        - application/json
      parameters:
        - name: networkId
          in: path
          description: ID of network to add user to.
          required: true
          type: integer
      responses:
        '200':
          description: OK
        '405':
          description: Invalid input
        '401':
          $ref: "#/responses/UnauthorizedError"
  '/user/leaveNetwork/{networkId}':
    delete:
      security:
        - basicAuth: []
      tags:
        - users
      summary: let user leave network.
      description: notice that no error will be thrown if user never joined the network in the first place.
      parameters:
        - in: path
          name: networkId
          required: true
          type: integer
          description: Id of network you want to leave.
      responses:
        '200':
          description: User 155 left network 544
        '401':
          $ref: "#/responses/UnauthorizedError"
  /user/users:
    get:
      tags:
        - users
      summary: Get users.
      description: ''
      operationId: getUsers
      produces:
        - application/json
      parameters:
        - in: query
          name: count
          type: integer
          description: |
            The number of results to return.  Between 1 and 100.
        - in: query
          name: max_id
          type: integer
          description: |
            The maximum ID, inclusive, to return data for.
        - in: query
          name: near_location
          type: string
          description: |
            A comma-separated list of country_id, region_id, and city_id, in that order.
        - in: query
          name: from_location
          type: string
          description: |
             A comma-separated list of country_id, region_id, and city_id, in that order.
        - in: query
          name: language
          type: string
          description: |
             Name of language.
      responses:
        '200':
          description: A list of users matching filter.
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        '405':
          description: Invalid input
    post:
      tags:
        - users
      summary: Create a new user
      description: ''
      operationId: createUser
      consumes:
        - application/json
      parameters:
        - in: body
          name: user
          description: Created user object
          required: true
          schema:
            $ref: '#/definitions/NewUser'
      responses:
        '200':
          description: OK
    put:
      tags:
        - users
      summary: Update a user
      description: Uses the ID given in the JSON object to determine
                   what user tuple to modify, then updates the rest
                   of the fields with the JSON values given. Note
                   that it is impossible to update a user id.
      operationId: updateUser
      consumes:
        - application/json
      parameters:
        - in: body
          name: body
          description: User object with which to update
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Network"
...
  '/event/{eventId}':
    get:
      tags:
        - events
      summary: Get an event.
      description: ''
      operationId: getEvent
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path
          description: ID of event to return.
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: Ok
          schema:
            $ref: '#/definitions/Event'
        '405':
          description: Invalid input
  '/event/{eventId}/reg':
    get:
      tags:
        - events
      summary: Get list of registration JSONs for an event.
      description: ''
      operationId: getEventAttendance
      produces:
        - application/json
      parameters:
        - name: eventId
          in: path
          description: ID of event to fetch attendees.
          required: true
          type: integer
          format: int64
        - in: query
          name: count
          type: integer
          description: |
            The number of results to return.  Between 1 and 100.
        - in: query
          name: max_register_date
          type: string
          format: timestamp
          description: |
            The latest date of registration, inclusive, to return data for.
      responses:
        '200':
          description: Ok
          schema:
            type: array
            items:
              $ref: '#/definitions/EventRegistration'
        '405':
          description: Invalid input
  /event/new:
    post:
      security:
            - basicAuth: []
      tags:
        - events
      summary: Create an event.
      description: ''
      operationId: newEvent
      consumes:
        - application/json
      parameters:
        - in: body
          name: event
          description: The event specification.
          required: true
          schema:
            $ref: '#/definitions/NewEvent'
      responses:
        '200':
          description: Ok
        '405':
          description: Invalid input
    put:
      security:
          - basicAuth: []
      tags:
        - events
      summary: Update an event.
      description: ''
      operationId: updateEvent
      consumes:
        - application/json
      parameters:
        - in: body
          name: event
          description: The event specification.
          required: true
          schema:
            $ref: '#/definitions/Event'
      responses:
        '200':
          description: Ok
        '405':
          description: Invalid input
...
responses:
  UnauthorizedError:
    description: Authentication information is missing or invalid
definitions:
...
  User:
    type: object
    required:
      - id
      - first_name
      - last_name
      - email
      - username
      - role
      - act_code
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      register_date:
        type: string
        format: timestamp
      role:
        type: integer
        format: int32
      last_login:
        type: string
        format: timestamp
      gender:
        type: string
      about_me:
        type: string
      confirmed:
        type: boolean
      act_code:
        type: string
        description: Not sure what this is.
      img_link:
        type: string
        description: URL of image
      fp_code:
        type: string
        description: Not sure what this is.
  NewUser:
    description: The fields needed to create a new user.
    type: object
    properties:
      username:
        type: string
      password:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      role:
        type: integer
        format: int32
      img_link:
        type: string
      about_me:
        type: string
      gender:
        type: string
  Network:
    type: object
    required:
      - id
      - location_cur
    properties:
      id:
        type: integer
        format: bigint20
      city_cur:
        type: string
      id_city_cur:
        type: integer
        format: bigint20
      region_cur:
        type: string
      id_region_cur:
        type: integer
        format: bigint2
      country_cur:
        type: string
      id_country_cur:
        type: integer
        format: bigint20
      city_origin:
        type: string
      id_city_origin:
        type: integer
        format: bigint20
      region_origin:
        type: string
      id_region_origin:
        type: integer
        format: bigint20
      country_origin:
        type: string
      id_country_origin:
        type: integer
        format: bigint20
      language_origin:
        type: string
      id_language_origin:
        type: integer
        format: bigint20
      network_class:
        type: string
        description: |
         _l - language
          cc - city
          rc - region
          co - country
      date_added:
        type: string
        format: timestamp
      img_link:
        type: string
      twitter_query_level:
        type: string
...
  NetworkWithJoin:
    description: |
      A Network with the join_date of the user for which this network
      was invoked for, e.g., the /user/{userId}/networks call.
    type: object
    required:
      - id
      - location_cur
    properties:
      id:
        type: integer
        format: bigint20
      city_cur:
        type: string
      id_city_cur:
        type: integer
        format: bigint20
      region_cur:
        type: string
      id_region_cur:
        type: integer
        format: bigint2
      country_cur:
        type: string
      id_country_cur:
        type: integer
        format: bigint20
      city_origin:
        type: string
      id_city_origin:
        type: integer
        format: bigint20
      region_origin:
        type: string
      id_region_origin:
        type: integer
        format: bigint20
      country_origin:
        type: string
      id_country_origin:
        type: integer
        format: bigint20
      language_origin:
        type: string
      id_language_origin:
        type: integer
        format: bigint20
      network_class:
        type: string
        description: |
         _l - language
          cc - city
          rc - region
          co - country
      date_added:
        type: string
        format: timestamp
      img_link:
        type: string
      twitter_query_level:
        type: string
      join_date:
        format: timestamp
...
  NewEvent:
    description: The fields needed to create a new event.
    type: object
    required:
      - id
    properties:
      id_network:
        type: integer
        format: int64
      id_host:
        type: integer
        format: int64
      event_date:
        type: string
        format: date-time
      title:
        type: string
      address_1:
        type: string
      address_2:
        type: string
      country:
        type: string
      city:
        type: string
      region:
        type: string
      description:
        type: string
  Event:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
      id_network:
        type: integer
        format: int64
      id_host:
        type: integer
        format: int64
      date_created:
        type: string
        format: timestamp
      event_date:
        type: string
        format: date-time
      title:
        type: string
      address_1:
        type: string
      address_2:
        type: string
      country:
        type: string
      city:
        type: string
      region:
        type: string
      description:
        type: string
  EventRegistration:
    type: object
    required:
      - id_guest
      - id_event
    properties:
      id_guest:
        type: integer
        format: int64
      id_event:
        type: integer
        format: int64
      date_registered:
        type: string
        format: timestamp
      job:
        type: string
  Post:
    type: object
    required:
      - id
    properties:
      id:
        type: integer
        format: int64
      id_user:
        type: integer
        format: int64
      id_network:
        type: integer
        format: int64
      post_date:
        type: string
        format: timestamp
      post_text:
        type: string
      post_class:
        type: integer
        description: Not sure what this is.
      post_original:
        type: string
        description: Not sure what this is.
      vid_link:
        type: string
      img_link:
        type: string
...